python_port/pyproject.toml
[project]
name = "game4s-python"
version = "0.1.0"
description = "Python rewrite of game4s with pygame client and FastAPI multiplayer API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "redis>=5.0.0",
  "pygame-ce>=2.5.7",
  "httpx>=0.27.0"
]

[tool.uv]
package = true
